requirements.txt
certifi==2020.12.5
cloudinary==1.24.0
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
django-heroku==0.3.1
djangorestframework==3.10.3
geographiclib==1.50
geopy==1.20.0
gunicorn==19.9.0
mpu==0.20.0
psycopg2==2.8.4
python-googlegeocoder==0.4.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
urllib3==1.26.4
whitenoise==4.1.4
